services/data_processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
xlsxwriter==3.1.9
aiofiles==23.2.0
httpx==0.25.2
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1